E-15: extract ConstructorBodyEmitter from ReflectionMetadataEmitter (#1361)
Extracts the constructor-body-bytes emission band (~793 lines) off the
ReflectionMetadataEmitter root into a new internal sealed
ConstructorBodyEmitter, following the E-14 back-reference idiom.

Methods moved verbatim (mechanical reference rewrites only):
  - EmitStaticConstructorBodyBytes (TypeDefEmitter callback)
  - EmitInterfaceStaticConstructor + EmitInterfaceStaticConstructorBodyBytes
    (invoked directly from EmitCore)
  - EmitStaticConstructorBodyFromBlock (shared .cctor core)
  - EmitClassDefaultConstructorBodyBytes (callback)
  - EmitClassPrimaryConstructorBodyBytes (callback)
  - EmitClassConstructorWithBaseInitializerBodyBytes (callback)
  - EmitClassConstructorWithBodyBodyBytes (callback)
  - EmitClassDeinitializerBodyBytes (callback)
  - the shared instance-field-initializer statement builders they call
    (NeedsInstanceFieldInitializerStatements, HasNestedStringField,
    BuildInstanceFieldInitializerStatements,
    AppendNestedStructStringDefaultAssignments)

Wiring:
  - RME constructs ConstructorBodyEmitter in EmitCore right after
    InterfaceImplEmitter, injecting the RME back-reference and the
    CustomAttributeEncoder.NextParameterHandle callback (used only by the
    interface-.cctor MethodDef's parameter-list anchor).
  - The TypeDefEmitter construction site rebinds its five *BodyBytes
    method-group args from this.X to this.ctorBodies.X.
  - The EmitInterfaceStaticConstructor call in EmitCore rebinds to
    this.ctorBodies.EmitInterfaceStaticConstructor.

State move (currentStaticConstructorOwner -> EmitContext):
  - The mutable Symbol field currentStaticConstructorOwner moved off RME
    onto EmitContext as a public auto-property CurrentStaticConstructorOwner
    { get; set; } (defaults to null, same as the field). Reference sites
    updated:
      * RME field declaration + its comment block (removed)
      * EmitStaticConstructorBodyBytes save/set/restore (3 sites) -> moved
        file, now emitCtx.CurrentStaticConstructorOwner
      * EmitInterfaceStaticConstructorBodyBytes save/set/restore (3 sites)
        -> moved file, now emitCtx.CurrentStaticConstructorOwner
      * ReflectionMetadataEmitter.IsStaticFieldAddressLegalHere read site
        -> this.emitCtx.CurrentStaticConstructorOwner (stays on root;
        reached from MethodBodyEmitter via outer.IsStaticFieldAddressLegalHere)

Widenings:
  - Nested MethodBodyEmitSession scaffold: private sealed -> internal sealed
    so ConstructorBodyEmitter can construct it as
    ReflectionMetadataEmitter.MethodBodyEmitSession(outer, il). It still has
    two other in-RME construction sites (BuildMoveNextBodyBytes, EmitFunction),
    so it stays on the root.
  - NeedsInstanceFieldInitializerStatements (already internal static) is
    called from TypeDefEmitter; its two call sites there were re-pointed to
    ConstructorBodyEmitter.NeedsInstanceFieldInitializerStatements (no
    forwarder needed — only in-assembly callers). ResolveFieldToken,
    TryGetPrimaryCtorTargetField, IsValueTypeSymbol,
    IsUserGenericTypeReference were already internal.

No forwarders retained: every external caller (TypeDefEmitter wiring +
EmitCore) was re-pointed directly.

Verification: dotnet build GSharp.sln -c Release -graph -> 0 warnings,
0 errors. PE byte-diff over samples/*.gs: 119/119 emitted PEs byte-identical
before/after (3 expected GsharpExtensions* baseline failures skipped). The
mutable-state move preserved emission order and values exactly.

RME 12031 -> 11251 lines; ConstructorBodyEmitter 855 lines; EmitContext
244 -> 259 lines.
